Moved Util class and added a method
