+++ src/test/java/edu/kit/ipd/dbis/org/jgrapht/additions/Util.java
@@ -1,0 +1,43 @@
+package edu.kit.ipd.dbis.org.jgrapht.additions;
+
+import edu.kit.ipd.dbis.org.jgrapht.additions.graph.PropertyGraph;
+
+public class Util {
+	public static PropertyGraph<Integer, Integer> createCompleteGraph(int numberOfVertices) {
+		PropertyGraph<Integer, Integer> graph = new PropertyGraph<>();
+		for (int i = 0; i < numberOfVertices; i++) {
+			graph.addVertex(i);
+		}
+		for (int i = 0; i < numberOfVertices; i++) {
+			for (int j = 0; j < numberOfVertices; j++) {
+				if (j != i && !graph.containsEdge(graph.getEdgeFactory().createEdge(j, i))) {
+					graph.addEdge(i, j);
+
+				}
+			}
+		}
+		return graph;
+	}
+
+	public static PropertyGraph generateSimpleTestGraph() {
+		PropertyGraph graph = new PropertyGraph();
+		graph.addVertex("a");
+		graph.addVertex("b");
+		graph.addVertex("c");
+		graph.addVertex("d");
+		graph.addVertex("e");
+		graph.addVertex("f");
+		graph.addVertex("g");
+		graph.addEdge("d", "a");
+		graph.addEdge("d", "e");
+		graph.addEdge("d", "b");
+		graph.addEdge("a", "f");
+		graph.addEdge("e", "f");
+		graph.addEdge("e", "b");
+		graph.addEdge("e", "g");
+		graph.addEdge("e", "c");
+		graph.addEdge("g", "c");
+		graph.addEdge("a", "g");
+		return graph;
+	}
+}
